+++ cmd/test.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"bytes"
 	"fmt"
+	"github.com/moshloop/platform-cli/pkg/phases/kubeadm"
 	"io/ioutil"
 	"os"
 	"path"
@@ -10,7 +11,6 @@ import (
 	"time"
 
 	"github.com/flanksource/commons/console"
-	"github.com/moshloop/platform-cli/pkg/phases/audit"
 	"github.com/moshloop/platform-cli/pkg/phases/base"
 	"github.com/moshloop/platform-cli/pkg/phases/configmapreloader"
 	"github.com/moshloop/platform-cli/pkg/phases/consul"
@@ -122,7 +122,7 @@ func init() {
 	}
 
 	tests := map[string]TestFn{
-		"audit":              audit.Test,
+		"audit":              kubeadm.Test,
 		"base":               base.Test,
 		"configmap-reloader": configmapreloader.Test,
 		"consul":             consul.Test,
